requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-tavily
langgraph
streamlit
python-dotenv
matplotlib
requests